+++ MABEngine/src/MABEngine/Inputs/Input.h
@@ -16,16 +16,16 @@ namespace MABEngine {
 
 			inline static bool IsMouseButtonPressed(MABMouseButton buttonCode) { return s_Instance->IsMouseButtonPressedImp(buttonCode); }
 			inline static std::pair<double, double> GetMousePos() { return s_Instance->GetMousePosImp(); }
-			inline static float GetMousePosX() { return s_Instance->GetMousePosXImp(); }
-			inline static float GetMousePosY() { return s_Instance->GetMousePosYImp(); }
+			inline static double GetMousePosX() { return s_Instance->GetMousePosXImp(); }
+			inline static double GetMousePosY() { return s_Instance->GetMousePosYImp(); }
 		protected:
 			virtual bool IsKeyPressedImp(MABKeyboardCode keyCode) = 0;
 			virtual bool IsKeyRepeatImp(MABKeyboardCode keyCode) = 0;
 
 			virtual bool IsMouseButtonPressedImp(MABMouseButton buttonCode) = 0;
 			virtual std::pair<double, double> GetMousePosImp() = 0;
-			virtual float GetMousePosXImp() = 0;
-			virtual float GetMousePosYImp() = 0;
+			virtual double GetMousePosXImp() = 0;
+			virtual double GetMousePosYImp() = 0;
 		private:
 			static Input* s_Instance;
 		};
